Chore: Run clippy

+++ src/fflonk/checkpairing_with_c_wi.rs
@@ -57,11 +57,11 @@ mod test {
                 r.pow([3_u32.pow(delta as u32).to_u64().unwrap()])
             };
             if d == cc {
-                (h, r) = (h * c, r * c.pow([3 as u64]));
+                (h, r) = (h * c, r * c.pow([3_u64]));
             } else if d == cc.pow([2_u64]) {
-                (h, r) = (h * c.pow([2 as u64]), r * c.pow([3 as u64]).pow([2 as u64]));
+                (h, r) = (h * c.pow([2_u64]), r * c.pow([3_u64]).pow([2_u64]));
             }
-            c = c.pow([3 as u64])
+            c = c.pow([3_u64])
         }
 
         // recover cubic root of a
@@ -70,7 +70,7 @@ mod test {
             r = r.inverse().unwrap();
         }
 
-        assert_eq!(r.pow([3 as u64]), a);
+        assert_eq!(r.pow([3_u64]), a);
         r
     }
